# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Thoth API",
    "version": "0.1.0",
    "description": "Live OSINT layers, intelligence, watches and on-demand lookups. See docs/reference/api.md for conventions (authentication, list and error shapes, the event object).",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4000",
      "description": "API, direct (host only)"
    },
    {
      "url": "/",
      "description": "Through the app (access token when gated)"
    }
  ],
...
  "paths": {
...
    "/api/stats": {
      "get": {
        "operationId": "getStats",
        "tags": [
          "System"
        ],
        "summary": "Row counts per layer",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/versions": {
      "get": {
        "operationId": "getVersions",
        "tags": [
          "System"
        ],
        "summary": "Layer versions; poll to find what changed",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/routes": {
      "get": {
        "operationId": "getRoutes",
        "tags": [
          "System"
        ],
        "summary": "Every registered route",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
    "/api/layers/{layer}": {
      "get": {
        "operationId": "getLayersLayer",
        "tags": [
          "Layers"
        ],
        "summary": "Layer slice",
        "description": "Without `z`: the newest 500 events. With `z`: a viewport slice with spatial sampling, plus matched, limit and truncated.",
        "parameters": [
          {
            "name": "layer",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "z",
            "in": "query",
            "required": false,
            "description": "Map zoom. With it, the response is a viewport slice.",
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 24
            }
          },
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "description": "Viewport `w,s,e,n` in degrees; w > e crosses the antimeridian.",
            "schema": {
              "type": "string",
              "maxLength": 120
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/layers/{layer}/history": {
      "get": {
        "operationId": "getLayersLayerHistory",
        "tags": [
          "Layers"
        ],
        "summary": "Event counts per time bucket",
        "parameters": [
          {
            "name": "layer",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bucket",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "hour",
                "day"
              ],
              "default": "day"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
    "/api/brief": {
      "get": {
        "operationId": "getBrief",
        "tags": [
          "Intelligence"
        ],
        "summary": "Deterministic brief: counts and top items",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/event": {
      "get": {
        "operationId": "getEvent",
        "tags": [
          "Intelligence"
        ],
        "summary": "One event with its provenance",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "Event id",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/alerts": {
      "get": {
        "operationId": "getAlerts",
        "tags": [
          "Intelligence"
        ],
        "summary": "Critical and watch events in a window",
        "parameters": [
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 168,
              "default": 24
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/dossier": {
      "get": {
        "operationId": "getDossier",
        "tags": [
          "Intelligence"
        ],
        "summary": "Everything near a point",
        "description": "Events by layer, geo context and a threat score.",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "minimum": -90,
              "maximum": 90
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "description": "Longitude (`lng` is accepted too).",
            "schema": {
              "type": "number",
              "minimum": -180,
              "maximum": 180
            }
          },
          {
            "name": "radius_km",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "ok": {
            "const": false
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "ok",
          "error"
        ]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid parameters",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "RateLimited": {
        "description": "Rate limited: retry after `Retry-After` seconds",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerWriteKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API_WRITE_KEY, for POST and DELETE"
      },
      "headerWriteKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Thoth-Key"
      }
    }
  }
}